requirements.txt
instabot
transformers
textblob
pillow
torch
torchvision